move the LocationTypeFilter into the map-nav-bar directory
