+++ src/app-pages/map/components/map-nav-bar/MapNavBar.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import SearchBox from "../../../../app-common/inputs/SearchBox";
 import Dropdown from "../../../../app-common/inputs/Dropdown";
 import LocationTree from "./LocationTree";
-import LocationTypeFilter from "../LocationTypeFilter";
+import LocationTypeFilter from "./LocationTypeFilter";
 import "./mapNavBar.scss";
 
 const MapNavBar = () => {
